Initial commit - GitMonitor
GitMonitor is a cog developed for QCAdmin, but could be loaded by any discord bot, provided that the bot owner has access to the GitHub API.

Since QCAdmin is a dockerized app, a GitHub api key with sufficient permissions to read the commit history of a repository will need to be provided with the Docker run.

The bot will need access to a config directory where it will manage the cog configuration.

Some of the primary methods of the cog may change significantly as this is tested in the dev branch.