Admin Logs command
Since the memory logger is not working as expected, wrote a simple function to return the contents of the log file generated by our logger instance. Argument can be provided for the number of lines from the log which should be returned.

As part of the admin commands group, the command can be called with /admin logs