Restructure and Changes for Docker
I've adapted this version of the cog from the functioning cog running in QCAdmin Dev. This version supports containerization expecting the github token to be passed with the docker run. The first time the cog is loaded a json configuration should be created in the mounted docker storage volume. Added commands to change the commit check loop time.