initial commit for marble fans
A lot of fleshing out is required specifically for the actual code to scrape the wiki table data with beautifulsoup. This initial commit includes the necessary imports, url, channel and token dependencies,  as well as the overall discord.py commands structure based on interactions with discord.py slash commands.

Will load code as is to the dev bot instance to dev/test effective wiki results scraper.